refactor(chat): extract bubble components and helpers from Conversations.tsx
Split the 1,692-line Conversations.tsx by lifting presentational bubbles and
pure helpers into app/src/pages/conversations/. Conversations.tsx now 1,394
lines; semantically identical (verified by Codex + Gemini review).

- components/AgentMessageBubble.tsx (+BubbleMarkdown)
- components/ToolTimelineBlock.tsx
- components/LimitPill.tsx
- utils/format.ts (AgentBubblePosition, formatRelativeTime, formatResetTime,
  getInlineCompletionSuffix, buildAcceptedInlineCompletion)

Typecheck clean, 464 tests pass.
